Add feature: expand / collapse all results for project search scope
When I change search scope to project, there are too many files and
comments in the result panel. So, I add this feature for finding the
file and comment more easier. Now it works fine for me.
The next feature I want to add is filtering the result by tag. There are
many software which has this feature. such as Visual Studio, Eclipse.
What do you think @mikaeljorhult? Any ideas?